+++ lib/petProfile_ownerProfile.dart
@@ -83,7 +83,7 @@ class pet extends StatelessWidget {
                 builder: (context, snapshot) {
                   if (!snapshot.hasData) return const Text('loading');
                   return ListView.builder(
-                    primary: false,
+                    physics: const NeverScrollableScrollPhysics(),
                     itemCount: snapshot.data!.docs.length,
                     itemBuilder: (context, index) =>
                     //card pets method
@@ -95,7 +95,7 @@ class pet extends StatelessWidget {
 
 
           Container(
-            padding: EdgeInsets.only(left:MediaQuery.of(context).size.width * 0.1,top: 200, right: 25),
+            padding: EdgeInsets.only(left:MediaQuery.of(context).size.width * 0.1,top: 180, right: 35),
             height: 1000,
 
             child: StreamBuilder<QuerySnapshot>(
@@ -104,9 +104,7 @@ class pet extends StatelessWidget {
                 builder: (context, snapshot) {
                   if (!snapshot.hasData) return const Text('loading');
                   return ListView.builder(
-                    scrollDirection: Axis.vertical,
-                    shrinkWrap: true,
-                    physics: const ClampingScrollPhysics(),
+                    physics: const NeverScrollableScrollPhysics(),
                     itemCount: snapshot.data!.docs.length,
                     itemBuilder: (context, index) =>
                     //card pets method
@@ -149,12 +147,13 @@ class pet extends StatelessWidget {
 
                   child: ListTile(
                     title: Text(
-                        "name:  " + document['name'] + "\nspecies:  " +
-                            document['species'] + "\ngender:  " +
-                            document['gender'] + "\nbirth date:  " +
+                        "\nName:  " + document['name'] + "\nSpecies:  " +
+                            document['species'] + "\nGender:  " +
+                            document['gender'] + "\nBirth date:  " +
                             document['birthDate'] , style: petCardTitleStyle),
 
                   ),),
+
 
                 Container(
                   margin: EdgeInsets.only(left: 15,right: 15,bottom: 40),
@@ -262,17 +261,20 @@ class pet extends StatelessWidget {
           actions: <Widget>[
             FlatButton(
               child: Text("YES"),
-              onPressed: () {
-                document.reference.delete().then((_){
+    onPressed: () async {
+    FirebaseFirestore firestoreInstance= FirebaseFirestore.instance;
 
-                  Navigator.pop(context, true);
-                }   );
+    QuerySnapshot<Map<String, dynamic>> snapshot = await firestoreInstance
+        .collection('appointment')
+        .where('petID', isEqualTo: document.id)
+        .get();
 
-                // Navigator.push(context,MaterialPageRoute(builder: (_) =>MyPets())) .catchError((error) => print('Delete failed: $error'));;
-                //Put your code here which you want to execute on Yes button click.
-
-              },
-            ),
+    if(snapshot.docs.length==0){
+    document.reference.delete().then((_) {
+    Navigator.pop(context, true);
+    });}
+    else{ Navigator.pop(context, false); }
+    }),
 
 
             FlatButton(
